+++ env.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include "types.h"
+
 // TODO something better than linear search through a linked list
 typedef struct entry
 {
   char *key;
-  void *value;
+  sexpr *value;
   struct entry *next;
 } entry;
 
@@ -14,6 +16,6 @@ typedef struct environment
   struct env *parent;
 } environment;
 
-void *env_get (environment *env, const char *key);
-void env_set (environment *env, const char *key, void *e);
+sexpr *env_get (environment *env, const char *key);
+void env_set (environment *env, const char *key, sexpr *e);
 // TODO delete?
